% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newrat_function.R
\name{newton.raphson}
\alias{newton.raphson}
\title{A newton raphson Function}
\usage{
newton.raphson(f, x0, tol = 1e-05, n = 1000)
}
\description{
This function allows you to calculate zeros of different functions
}
\examples{
newton.raphson()
}
\keyword{newton}
\keyword{raphson}
\keyword{roots}
